refactor: move notes in tera-project.vue to its own component (#1118)
