+++ packages/client/hmi-client/src/page/project/tera-project.vue
@@ -13,7 +13,7 @@
 					:tabs="tabs"
 					:active-tab="openedAssetRoute"
 					@open-asset="openAssetFromSidebar"
-					@click="getAndPopulateAnnotations()"
+					@close-tab="removeClosedTab"
 					@remove-asset="removeAsset"
 				/>
 			</template>
@@ -28,7 +28,6 @@
 					:loading-tab-index="loadingTabIndex"
 					@close-tab="removeClosedTab"
 					@select-tab="openAsset"
-					@click="getAndPopulateAnnotations()"
 				/>
 				<tera-project-page
 					v-if="project"
@@ -75,137 +74,13 @@
 			</SplitterPanel>
 		</Splitter>
 		<tera-slider-panel
-			class="slider"
+			v-model:is-open="isNotesSliderOpen"
 			content-width="240px"
 			direction="right"
 			header="Notes"
-			v-model:is-open="isNotesSliderOpen"
-			@click="getAndPopulateAnnotations()"
 		>
 			<template v-slot:content>
-				<section class="annotation-panel-container">
-					<div v-for="(annotation, idx) of annotations" :key="idx">
-						<div
-							v-if="isEditingNote && idx === selectedNoteIndex"
-							class="annotation-input-container"
-						>
-							<div class="annotation-header">
-								<Dropdown
-									placeholder="Unassigned"
-									class="p-button p-button-text notes-dropdown-button"
-									:options="noteOptions"
-									v-model="selectedNoteSection[idx]"
-								/>
-							</div>
-							<Textarea
-								v-model="annotation.content"
-								ref="annotationTextInput"
-								rows="5"
-								cols="30"
-								aria-labelledby="annotation"
-							/>
-							<div class="save-cancel-buttons">
-								<Button
-									@click="isEditingNote = false"
-									label="Cancel"
-									class="p-button p-button-secondary"
-									size="small"
-								/>
-								<Button
-									@click="
-										updateNote();
-										isEditingNote = false;
-									"
-									label=" Save"
-									class="p-button"
-									size="small"
-								/>
-							</div>
-						</div>
-						<div v-else>
-							<div class="annotation-header">
-								<!-- TODO: Dropdown menu is for selecting which section to assign the note to: Unassigned, Abstract, Methods, etc. -->
-								<Dropdown
-									disabled
-									placeholder="Unassigned"
-									class="p-button p-button-text notes-dropdown-button"
-									:options="noteOptions"
-									v-model="selectedNoteSection[idx]"
-								/>
-								<!-- TODO: Ellipsis button should open a menu with options to: Edit note & Delete note -->
-								<Button
-									icon="pi pi-ellipsis-v"
-									class="p-button-rounded p-button-secondary"
-									@click="
-										(event) => {
-											toggleAnnotationMenu(event);
-											selectedNoteIndex = idx;
-										}
-									"
-								/>
-							</div>
-							<div>
-								<p>{{ annotation.content }}</p>
-								<div class="annotation-author-date">
-									<div>
-										{{ formatAuthorTimestamp(annotation.username, annotation.timestampMillis) }}
-									</div>
-								</div>
-							</div>
-						</div>
-					</div>
-					<Menu
-						ref="annotationMenu"
-						:model="annotationMenuItems"
-						:popup="true"
-						@hide="onHide"
-						@click.stop
-					/>
-				</section>
-				<div class="annotation-input-box">
-					<div v-if="isAnnotationInputOpen" class="annotation-input-container">
-						<div class="annotation-header">
-							<Dropdown
-								placeholder="Unassigned"
-								class="p-button p-button-text notes-dropdown-button"
-								:options="noteOptions"
-								v-model="newNoteSection"
-							/>
-						</div>
-						<Textarea
-							v-model="annotationContent"
-							ref="annotationTextInput"
-							rows="5"
-							cols="30"
-							aria-labelledby="annotation"
-						/>
-						<div class="save-cancel-buttons">
-							<Button
-								@click="toggleAnnotationInput()"
-								label="Cancel"
-								class="p-button p-button-secondary"
-								size="small"
-							/>
-							<Button
-								@click="
-									addNote();
-									toggleAnnotationInput();
-								"
-								label=" Save"
-								class="p-button"
-								size="small"
-							/>
-						</div>
-					</div>
-					<div v-else>
-						<Button
-							@click="toggleAnnotationInput()"
-							icon="pi pi-plus"
-							label="Add note"
-							class="p-button-text p-button-flat"
-						/>
-					</div>
-				</div>
+				<tera-notes-sidebar :asset-id="assetId" :page-type="pageType" />
 			</template>
 		</tera-slider-panel>
 	</main>
@@ -218,26 +93,16 @@ import { isEmpty, isEqual } from 'lodash';
 import TeraModelWorkflowWrapper from '@/components/workflow/tera-model-workflow-wrapper.vue';
 import TeraDatasetWorkflowWrapper from '@/components/workflow/tera-dataset-workflow-wrapper.vue';
 import { WorkflowNode, WorkflowOperationTypes } from '@/types/workflow';
-import Button from 'primevue/button';
-import Dropdown from 'primevue/dropdown';
-import Textarea from 'primevue/textarea';
 import TeraSliderPanel from '@/components/widgets/tera-slider-panel.vue';
 import TeraTabGroup from '@/components/widgets/tera-tab-group.vue';
 import TeraResourceSidebar from '@/page/project/components/tera-resource-sidebar.vue';
+import TeraNotesSidebar from '@/page/project/components/tera-notes-sidebar.vue';
 import { RouteName } from '@/router/routes';
 import * as ProjectService from '@/services/project';
 import { useTabStore } from '@/stores/tabs';
-import { Tab, Annotation } from '@/types/common';
+import { Tab } from '@/types/common';
 import { IProject, ProjectAssetTypes, ProjectPages, isProjectAssetTypes } from '@/types/Project';
 import { logger } from '@/utils/logger';
-import { formatDdMmmYyyy, formatLocalTime, isDateToday } from '@/utils/date';
-import {
-	createAnnotation,
-	deleteAnnotation,
-	getAnnotations,
-	updateAnnotation
-} from '@/services/models/annotations';
-import Menu from 'primevue/menu';
 import Splitter from 'primevue/splitter';
 import SplitterPanel from 'primevue/splitterpanel';
 import TeraCalibration from '@/components/workflow/tera-calibration.vue';
@@ -260,17 +125,6 @@ const tabStore = useTabStore();
 
 const router = useRouter();
 
-const isResourcesSliderOpen = ref(true);
-const isNotesSliderOpen = ref(false);
-const annotations = ref<Annotation[]>([]);
-const annotationContent = ref<string>('');
-const isAnnotationInputOpen = ref(false);
-const annotationMenu = ref();
-const menuOpenEvent = ref();
-const selectedNoteIndex = ref();
-const isEditingNote = ref(false);
-const isNoteDeletionConfirmation = ref(false);
-
 const workflowNode = ref<WorkflowNode | null>(null);
 // const workflowOperation = ref<string>('');
 
@@ -279,77 +133,8 @@ workflowEventBus.on('drilldown', (payload: any) => {
 	workflowNode.value = payload;
 });
 
-const noteDeletionConfirmationMenuItem = {
-	label: 'Are you sure?',
-	icon: '',
-	items: [
-		{
-			label: 'Yes, delete this note',
-			command: () => deleteNote()
-		}
-	]
-};
-const annotationMenuItems = ref([
-	{
-		label: 'Edit',
-		icon: 'pi pi-fw pi-file-edit',
-		command: () => {
-			isEditingNote.value = true;
-		}
-	},
-	{
-		label: 'Delete',
-		icon: 'pi pi-fw pi-trash',
-		command: () => {
-			if (!isNoteDeletionConfirmation.value) {
-				// @ts-ignore
-				annotationMenuItems.value.push(noteDeletionConfirmationMenuItem);
-				isNoteDeletionConfirmation.value = true;
-			}
-		}
-	}
-]);
-
-function onHide() {
-	if (isNoteDeletionConfirmation.value) {
-		annotationMenu.value.show(menuOpenEvent.value);
-		isNoteDeletionConfirmation.value = false;
-	} else if (annotationMenuItems.value.length > 2) {
-		annotationMenuItems.value.pop();
-	}
-}
-
-const toggleAnnotationMenu = (event) => {
-	// Fake object to allow the Menu component to not hide after clicking an item.
-	// Actually I am immediately showing it again after it automatically hides.
-	// I need to save and pass event.currentTarget in a ref to use when I want to manually show the menu.
-	// Kind of a hack/workaround due to the restrictive nature of this component.
-	menuOpenEvent.value = {
-		currentTarget: event.currentTarget
-	};
-	annotationMenu.value.toggle(event);
-};
-
-enum NoteSection {
-	Unassigned = 'Unassigned',
-	Abstract = 'Abstract',
-	Intro = 'Intro',
-	Methods = 'Methods',
-	Results = 'Results',
-	Discussion = 'Discussion',
-	References = 'References'
-}
-const noteOptions = ref([
-	NoteSection.Unassigned,
-	NoteSection.Abstract,
-	NoteSection.Intro,
-	NoteSection.Methods,
-	NoteSection.Results,
-	NoteSection.Discussion,
-	NoteSection.References
-]);
-const selectedNoteSection = ref<string[]>([]);
-const newNoteSection = ref();
+const isResourcesSliderOpen = ref(true);
+const isNotesSliderOpen = ref(false);
 
 // Associated with tab storage
 const projectContext = computed(() => props.project?.id.toString());
@@ -466,57 +251,6 @@ watch(
 tabStore.$subscribe(() => {
 	openAsset(tabStore.getActiveTabIndex(projectContext.value));
 });
-
-async function getAndPopulateAnnotations() {
-	if (props.assetId && props.pageType) {
-		annotations.value = await getAnnotations(props.assetId, props.pageType);
-		selectedNoteSection.value = annotations.value?.map((note) => note.section);
-	} else {
-		selectedNoteSection.value = [];
-	}
-}
-
-const addNote = async () => {
-	await createAnnotation(
-		newNoteSection.value,
-		annotationContent.value,
-		props.assetId,
-		props.pageType
-	);
-	annotationContent.value = '';
-	newNoteSection.value = NoteSection.Unassigned;
-	await getAndPopulateAnnotations();
-};
-
-async function updateNote() {
-	const noteToUpdate: Annotation = annotations.value[selectedNoteIndex.value];
-	const section =
-		selectedNoteSection.value.length >= selectedNoteIndex.value
-			? selectedNoteSection.value[selectedNoteIndex.value]
-			: null;
-	await updateAnnotation(noteToUpdate.id, section, noteToUpdate.content);
-	await getAndPopulateAnnotations();
-}
-
-async function deleteNote() {
-	const noteToDelete: Annotation = annotations.value[selectedNoteIndex.value];
-	await deleteAnnotation(noteToDelete.id);
-	await getAndPopulateAnnotations();
-}
-
-function toggleAnnotationInput() {
-	isAnnotationInputOpen.value = !isAnnotationInputOpen.value;
-	if (isAnnotationInputOpen.value === false) {
-		annotationContent.value = '';
-	}
-}
-
-function formatAuthorTimestamp(username, timestamp) {
-	if (isDateToday(timestamp)) {
-		return `${username} at ${formatLocalTime(timestamp)} today`;
-	}
-	return `${username} on ${formatDdMmmYyyy(timestamp)}`;
-}
 </script>
 
 <style scoped>
@@ -571,70 +305,5 @@ section,
 	display: inline-block;
 	overflow: hidden;
 	text-overflow: ellipsis;
-}
-
-.annotation-header {
-	display: flex;
-	justify-content: space-between;
-	height: 2rem;
-}
-
-.annotation-header .p-button.p-button-secondary {
-	background-color: var(--surface-section);
-}
-
-.annotation-panel-container {
-	display: flex;
-	gap: 16px;
-	padding: 0 16px 0 16px;
-}
-
-.notes-dropdown-button {
-	color: var(--text-color-subdued);
-	padding: 0rem;
-}
-
-.annotation-panel .p-dropdown:not(.p-disabled).p-focus {
-	box-shadow: none;
-	background-color: var(--surface-hover);
-}
-
-.annotation-author-date {
-	color: var(--text-color-subdued);
-	font-size: var(--font-caption);
-	display: flex;
-	justify-content: space-between;
-	gap: 0.5rem;
-	padding-top: 0.25rem;
-	padding-right: 1rem;
-}
-
-.annotation-input-container {
-	display: flex;
-	flex-direction: column;
-	gap: 8px;
-}
-
-.annotation-input-box {
-	padding: 16px 16px 8px 16px;
-}
-
-.annotation-input-box .p-inputtext {
-	border-color: var(--surface-border);
-	max-width: 100%;
-	min-width: 100%;
-}
-
-.annotation-input-box .p-inputtext:hover {
-	border-color: var(--primary-color) !important;
-}
-
-.save-cancel-buttons {
-	display: flex;
-	gap: 0.5rem;
-}
-
-.save-cancel-buttons > button {
-	flex: 1;
 }
 </style>
